Lots of function moved to scene, WinGL runs but doesn't render anything as need step is major overhaul of render states and binding (which aren't portable). Screen abstraction added but not fully used yet. However will allow true multi-monitor/window and stereo when ready Scene is now renderer independent CL disassociated from GL renderer for now

+++ src/libs/cl/context.h
@@ -15,7 +15,7 @@ namespace Cl {
 	public:
 		cl_context							getContext() { return context; }
 
-		int									getNumDevices() const { return devices.size(); }
+		int									getNumDevices() const { return (int) devices.size(); }
 		const cl_device_id*					getDevices() const { return &devices[0]; }
 		CmdQueuePtr							getCmdQueue( int deviceIndex ) { return cmdQueues[deviceIndex]; }
 		const DEVICE_EXTENSIONS_FLAGS*		getDeviceFlags() const { return &deviceFlags[0]; }
